+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.1.2
 - Added `/dump` `/time` `/ping` commands to the commands of the bot.
+- Added `/version` command and versioning of the bot.
 # 0.1.1
 - Added rate limit
 - Added subscription on many applications via 1 command